# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/confirm-email": {
      "post": {
        "operationId": "AuthController_confirmEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmEmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/library/add": {
      "post": {
        "operationId": "LibraryController_addToLibrary",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToLibraryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibrarySong"
                }
              }
            }
          }
        },
        "tags": [
          "Library"
        ]
      }
    },
    "/library/{songId}": {
      "delete": {
        "operationId": "LibraryController_removeFromLibrary",
        "parameters": [
          {
            "name": "songId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Library"
        ]
      }
    },
    "/library/like/{songId}": {
      "post": {
        "operationId": "LibraryController_toggleLike",
        "parameters": [
          {
            "name": "songId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Library"
        ]
      }
    },
    "/library/play/{id}": {
      "post": {
        "operationId": "LibraryController_recordPlay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Library"
        ]
      }
    },
...
    "/downloads/{songId}": {
      "post": {
        "operationId": "DownloadController_startDownload",
        "parameters": [
          {
            "name": "songId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartDownloadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Download"
        ]
      }
    },
    "/downloads/status/{jobId}": {
      "get": {
        "operationId": "DownloadController_getStatus",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Download"
        ]
      }
    },
    "/downloads/{jobId}": {
      "delete": {
        "operationId": "DownloadController_cancelDownload",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Download"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "AdminController_getAllUsers",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/search": {
      "get": {
        "operationId": "AdminController_searchUsers",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}": {
      "get": {
        "operationId": "AdminController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}/upgrade": {
      "put": {
        "operationId": "AdminController_upgradeToPremium",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpgradeToPremiumDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}/downgrade": {
      "put": {
        "operationId": "AdminController_downgradeToFree",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/comments": {
      "post": {
        "operationId": "CommentsController_createComment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
    "/comments/song/{songId}": {
      "get": {
        "operationId": "CommentsController_getSongComments",
        "parameters": [
          {
            "name": "songId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
    "/comments/{commentId}/replies": {
      "get": {
        "operationId": "CommentsController_getReplies",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
    "/comments/{id}": {
      "put": {
        "operationId": "CommentsController_updateComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        },
        "tags": [
          "Comments"
        ]
      },
      "delete": {
        "operationId": "CommentsController_deleteComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Your API",
    "description": "Your API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "SignUpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "username"
        ]
      },
      "ConfirmEmailDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "PlaylistSong": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "playlist": {
            "$ref": "#/components/schemas/Playlist"
          },
          "playlistId": {
            "type": "string"
          },
          "song": {
            "type": "object"
          },
          "songId": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "addedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "playlist",
          "playlistId",
          "song",
          "songId",
          "position",
          "addedAt"
        ]
      },
      "Playlist": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "source": {
            "enum": [
              "user",
              "spotify",
              "youtube"
            ],
            "type": "string"
          },
          "externalId": {
            "type": "string"
          },
          "coverImage": {
            "type": "string"
          },
          "isEditable": {
            "type": "boolean"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "string"
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaylistSong"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "source",
          "externalId",
          "coverImage",
          "isEditable",
          "user",
          "userId",
          "songs",
          "createdAt",
          "updatedAt"
        ]
      },
      "SearchHistory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "string"
          },
          "query": {
            "type": "string"
          },
          "filter": {
            "type": "string"
          },
          "searchedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user",
          "userId",
          "query",
          "filter",
          "searchedAt"
        ]
      },
      "Artist": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "lastFMLink": {
            "type": "string"
          },
          "externalListens": {
            "type": "number"
          },
          "mbid": {
            "type": "number"
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Song"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "songs",
          "createdAt",
          "updatedAt"
        ]
      },
      "Song": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "artistName": {
            "type": "string"
          },
          "albumName": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "albumCover": {
            "type": "string"
          },
          "spotifyId": {
            "type": "string"
          },
          "youtubeId": {
            "type": "string"
          },
          "externalUrl": {
            "type": "string"
          },
          "hasFlac": {
            "type": "boolean"
          },
          "slskPath": {
            "type": "string"
          },
          "artistId": {
            "type": "number"
          },
          "lastFMLink": {
            "type": "string"
          },
          "externalListens": {
            "type": "number"
          },
          "mbid": {
            "type": "string"
          },
          "downloadedPath": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "playCount": {
            "type": "number"
          },
          "likeCount": {
            "type": "number"
          },
          "commentCount": {
            "type": "number"
          },
          "repostCount": {
            "type": "number"
          },
          "librarySongs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibrarySong"
            }
          },
          "playlistSongs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaylistSong"
            }
          },
          "playHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayHistory"
            }
          },
          "artist": {
            "$ref": "#/components/schemas/Artist"
          },
          "relatedSongs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Song"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "artistName",
          "albumName",
          "duration",
          "albumCover",
          "spotifyId",
          "youtubeId",
          "externalUrl",
          "hasFlac",
          "slskPath",
          "artistId",
          "lastFMLink",
          "externalListens",
          "mbid",
          "downloadedPath",
          "metadata",
          "playCount",
          "likeCount",
          "commentCount",
          "repostCount",
          "librarySongs",
          "playlistSongs",
          "playHistory",
          "artist",
          "relatedSongs",
          "createdAt",
          "updatedAt"
        ]
      },
      "PlayHistory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "string"
          },
          "song": {
            "$ref": "#/components/schemas/Song"
          },
          "songId": {
            "type": "string"
          },
          "playedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user",
          "userId",
          "song",
          "songId",
          "playedAt"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "isPrivate": {
            "type": "boolean"
          },
          "googleId": {
            "type": "string"
          },
          "isEmailConfirmed": {
            "type": "boolean"
          },
          "emailConfirmToken": {
            "type": "string"
          },
          "resetPasswordToken": {
            "type": "string"
          },
          "apiKey": {
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string"
          },
          "subscriptionPlan": {
            "enum": [
              "free",
              "premium"
            ],
            "type": "string"
          },
          "subscriptionExpiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "swagz": {
            "type": "number"
          },
          "playlists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Playlist"
            }
          },
          "librarySongs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibrarySong"
            }
          },
          "searchHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchHistory"
            }
          },
          "playHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayHistory"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "password",
          "username",
          "bio",
          "avatarUrl",
          "isPrivate",
          "googleId",
          "isEmailConfirmed",
          "emailConfirmToken",
          "resetPasswordToken",
          "apiKey",
          "role",
          "subscriptionPlan",
          "subscriptionExpiresAt",
          "swagz",
          "playlists",
          "librarySongs",
          "searchHistory",
          "playHistory",
          "createdAt",
          "updatedAt"
        ]
      },
      "LibrarySong": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "string"
          },
          "song": {
            "$ref": "#/components/schemas/Song"
          },
          "songId": {
            "type": "string"
          },
          "isLiked": {
            "type": "boolean"
          },
          "isDownloaded": {
            "type": "boolean"
          },
          "addedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user",
          "userId",
          "song",
          "songId",
          "isLiked",
          "isDownloaded",
          "addedAt"
        ]
      },
      "AddToLibraryDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "album": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string"
          },
          "isLiked": {
            "type": "boolean"
          }
        },
        "required": [
          "title",
          "artist"
        ]
      },
...
      "StartDownloadDto": {
        "type": "object",
        "properties": {
          "preferFlac": {
            "type": "boolean"
          }
        }
      },
      "UpgradeToPremiumDto": {
        "type": "object",
        "properties": {
          "expiresAt": {
            "type": "string"
          }
        }
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "songId": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "parentCommentId": {
            "type": "string"
          }
        },
        "required": [
          "songId",
          "content"
        ]
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "string"
          },
          "song": {
            "$ref": "#/components/schemas/Song"
          },
          "songId": {
            "type": "string"
          },
          "parentCommentId": {
            "type": "string"
          },
          "parentComment": {
            "$ref": "#/components/schemas/Comment"
          },
          "isEdited": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "content",
          "user",
          "userId",
          "song",
          "songId",
          "parentCommentId",
          "parentComment",
          "isEdited",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          }
        },
        "required": [
          "content"
        ]
      },
...
    }
  }
}